Modifications to GUI
