+++ scripts/templates/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Random Poem Generator</title>
+    <title>DEEP FROST</title>
 
     <!-- CSS only -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css"
@@ -25,7 +25,7 @@
 <body>
     <div class="container">
         <div class="row">
-        <h1 class="mx-auto pt-2">Random Poem Generator</h1>
+        <h1 class="mx-auto pt-2">DEEP FROST</h1>
         </div>
         <hr>
         <div class="row pt-2">
@@ -48,7 +48,6 @@
         </div>
         <div class="col-7 ml-auto">
             {% if poem %}
-            <h5>The generated poem is:</h5><br>
             {% for line in poem %}
             <p>{{line|capitalize}}</p>
             {% endfor %}
